Split body and main templates
