+++ app/ui-master/client/body.html
@@ -1,0 +1,2 @@
+<body class="color-primary-font-color">
+</body>
