+++ app/ui-master/client/index.js
@@ -1,5 +1,5 @@
+import './body';
 import './loading';
 import './error.html';
 import './logoLayout.html';
-import './main.html';
 import './main';
