+++ app/ui-master/client/main.html
@@ -1,6 +1,3 @@
-<body class="color-primary-font-color">
-</body>
-
 <template name="main">
 	{{#if subsReady}}
 		{{#unless showSetupWizard}}
